Script updated with new number, description and common postconditions
